requirements.txt
pefile
macholib
pyelftools